backend/requirements.txt
pandas>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-google-genai>=0.0.3
langchain-community>=0.0.10
langchain-core>=0.1.0
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
google-ai-generativelanguage>=0.4.0
python-multipart>=0.0.6
pydantic>=2.0.0
tiktoken>=0.5.1
setuptools>=68.0.0
wheel>=0.41.0
numpy>=1.24.0
tqdm>=4.65.0
requests>=2.31.0
langgraph>=0.0.15
PyPDF2>=3.0.0
chardet>=5.0.0
llama-index
unstructured[all-docs]
tavily-python